refactor: use lowercase template filenames
